Fix content column breakpoints
